POSSIBLE FIX - see notes
Rewriting the algorithm seems to have, maybe, fixed the problem.
I still need to do more testing. However, testing it in the lab
gave results within 5% of where it should be. However, this
required adding a 35-ohm value to the measured resistance of the
potentiometer.

I'm not sure where this 35-ohm offset is coming from. Further testing
will be needed

NOTE: The YSI700 functionality has not been tested yet.